pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp/plan_execute", "mcp/llm"]

[project]
name = "assetopsbench-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "INSTRUCTIONS.md"
requires-python = ">=3.14"
dependencies = [
    "couchdb3>=2.0.2",
    "fastmcp>=2.14.5",
    "mcp[cli]>=1.26.0",
    "pendulum>=3.2.0",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
    "pyyaml>=6.0",
    "langchain-core>=0.3",
    "langchain-ibm>=0.3",
    "litellm>=1.0",
]

[project.scripts]
plan-execute = "plan_execute.cli:main"


[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "anyio>=4.9.0",
    "pytest-anyio>=0.0.0",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
pythonpath = ["mcp"]
filterwarnings = [
    "ignore:Core Pydantic V1 functionality:UserWarning",
]

